ion: Limit concurrency of workqueues freeing buffers asynchronously
Unbound workqueues do not participate in workqueue concurrency management,
and users of unbound workqueues are instead expected to handle concurrency
management themselves. This is problematic for ION since there's no limit
on the concurrency of buffer-free workers, meaning that they can starve
other SCHED_NORMAL processes which have the same nice value of zero or a
higher nice value.

The default number of workers allowed to run concurrently on *each* CPU in
an unbound workqueue's allowed CPUs is WQ_DFL_ACTIVE, which is a whopping
256 workers *per CPU*. This is way too high.

Reduce the number of concurrent buffer free workers to just one per CPU to
prevent them from starving other processes. Note that this cannot be done
by simply passing 1 to @max_active in alloc_workqueue() because doing so
with WQ_UNBOUND results in the workqueue becoming implicitly ordered and
thus only allowing a single worker to run at a time within the entire
workqueue. Instead, this is achieved using workqueue_set_max_active(),
which is the designated override for the implicit ordering quirk.

Also, remove WQ_CPU_INTENSIVE since it's redundant when WQ_UNBOUND is set.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>

Signed-off-by: munel <starscroch90@gmail.com>