arm64: allow LD_DEAD_CODE_DATA_ELIMINATION
- disable place-relative 32-bit relocations for DCE

DCE discards the evidently unused initcall pointers for integrated modules
when it's enabled. To prevent DCE from doing this, the initcall pointers
need to be C variables rather than inline assembly variables.

- arm64: Keep alternative-instruction sections

Otherwise DCE happily discards them and we're left with a kernel that
doesn't boot.

- arm64: Allow LD_DEAD_CODE_DATA_ELIMINATION to be selected

DCE is not problematic on arm64 and works out of the box. Let it be
used.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>

Signed-off-by: munel <starscroch90@gmail.com>